AROUNDMe identity server
This is one of several projects published at barnraiser.org.
Their description says, "the perfect solution for you to create identities for multiple people such as your family, a community or your company on the web."
This is a copy of the latest version aroundme_i_20080401.tar.gz found at http://barnraiser.org/aroundme.